internal/agentassets/agent_scripts/requirements.txt
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
argcomplete==3.6.2
backoff==2.2.1
beautifulsoup4==4.13.4
black==25.1.0
boto3==1.38.24
botocore==1.38.24
browser-use==0.2.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Cython==3.1.1
datamodel-code-generator==0.30.1
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.11.0
filetype==1.2.0
genson==1.3.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflect==7.5.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.42
markdownify==1.1.0
MarkupSafe==3.0.2
mem0ai==0.1.93
monotonic==1.6
more-itertools==10.7.0
mypy_extensions==1.1.0
numpy==2.2.6
ollama==0.4.9
openai==1.82.0
orjson==3.10.18
packaging==24.2
patchright==1.52.4
pathspec==0.12.1
platformdirs==4.3.8
playwright==1.52.0
portalocker==2.10.1
posthog==3.25.0
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyee==13.0.0
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.13.0
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
steel-sdk==0.3.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typeguard==4.4.2
typing_extensions==4.13.2
urllib3==2.4.0
uuid7==0.1.0
zstandard==0.23.0
